backend/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
psycopg2-binary
pgvector
python-multipart
python-jose[cryptography]
passlib
bcrypt==3.2.2
pydantic-settings
pydantic[email]
google-generativeai # For Gemini API
openai # For OpenAI API (optional, but good to have)
langchain
langchain-community
langchain-google-genai
python-dotenv
pypdf
chains